requirements.txt
Pandas
PyYAML
dill
Numpy
Matplotlib
seaborn
Scikit-Learn
Flask
gunicorn
statsmodels==0.14.1
astroid==2.5.6
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
nbformat==5.1.3
parso==0.8.2
patsy
pickleshare==0.7.5
Pillow==8.3.1
plotly
prompt-toolkit==3.0.18
Pygments==2.9.0
pylint==2.8.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil
pytz==2021.1
pyzmq
scipy
six
sklearn==0.0
tenacity==8.0.1
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.4.2
